Logging and Features
This is everything from v0.2.4-alpha including the logging functions.
logThis() can be accessed in all PHP parts of the system.

+++ delete.php
@@ -1,7 +1,3 @@
-<!-- Copyright GRAPHITE 2013 -->
-<!-- Grozinger P -->
-<!-- Formulate v1.0 -->
-
 <?php
 	session_start();
 ?>
@@ -14,13 +10,17 @@
 	</head>
 	<body>
 		<?php
-			echo "hi";
 			$data_con = mysqli_connect("localhost", 'root', 'root', 'formulae');
-			if (mysqli_query($data_con, "DELETE FROM formulae WHERE id='".$_GET["id"]."'")){
-				echo "Deleted";
-			}
-			else {
-				echo "Whooops... There was an error. Try again.";
+			$name = mysqli_query($data_con, "SELECT name FROM formulae WHERE id='".$_GET["id"]."'");
+			if (mysqli_num_rows($name) != 0){
+				if (mysqli_query($data_con, "DELETE FROM formulae WHERE id='".$_GET["id"]."'")){
+					echo "Deleted";
+					$name = mysqli_fetch_assoc($name);
+					logThis($name["name"]." was successfully deleted by ".$_SESSION["peep"].".");
+				}
+				else {
+					echo "Whooops... There was an error. Try again.";
+				}
 			}
 		?>
 		<script type="text/javascript">
